requirements.txt
Django==5.0.9
djangorestframework==3.15.0
django-rest-framework-simplejwt==5.3.0
drf-yasg==1.21.7
django-environ==0.11.2
celery==5.3.4
redis==5.0.1
Pillow==10.2.0